+++ tests/test_pytest_plugin.py
@@ -1,54 +1,10 @@
+from pathlib import Path
+
 import pytest
 
-CONFTEST = """
-import pytest
-
-
-@pytest.fixture(scope="session")
-def owner(accounts):
-    return accounts[0]
-
-
-@pytest.fixture(scope="session")
-def contract(project, owner):
-    c = project.TestContractVy.deploy(sender=owner)
-
-    # Show that contract transactions in fixtures appear in gas report
-    c.setNumber(999, sender=owner)
-
-    return c
-"""
-TEST_FILE = """
-def test_provider(networks):
-    # The default gets set in `ape-config.yaml`
-    assert networks.provider.name == "hardhat"
-
-def test_contract_interaction(owner, contract):
-    contract.setNumber(123, sender=owner)
-    assert contract.myNumber() == 123
-
-def test_transfer(accounts):
-    # Useful for seeing transfer gas
-    accounts[0].transfer(accounts[1], "100 gwei")
-
-def test_using_contract_with_same_name(owner, project):
-    contract = project.TestContractVy.deploy(sender=owner)
-    contract.setNumber(123, sender=owner)
-    assert contract.myNumber() == 123
-
-def test_two_contracts_with_same_symbol(owner, accounts, project):
-    # Tests against scenario when using 2 tokens with same symbol.
-    # There was almost a bug where the contract IDs clashed.
-    # This is to help prevent future bugs related to this.
-
-    receiver = accounts[1]
-    token_a = project.TokenA.deploy(sender=owner)
-    token_b = project.TokenB.deploy(sender=owner)
-    token_a.transfer(receiver, 5, sender=owner)
-    token_b.transfer(receiver, 6, sender=owner)
-    assert token_a.balanceOf(receiver) == 5
-    assert token_b.balanceOf(receiver) == 6
-"""
+BASE_DATA_PATH = Path(__file__).parent / "data" / "python"
+CONFTEST = (BASE_DATA_PATH / "pytest_test_conftest.py").read_text()
+TEST_FILE = (BASE_DATA_PATH / "pytest_tests").read_text()
 NUM_TESTS = len([x for x in TEST_FILE.split("\n") if x.startswith("def test_")])
 EXPECTED_GAS_REPORT = r"""
                    vyper_contract.json Gas
@@ -84,7 +40,7 @@ def ape_pytester(project, pytester):
     return pytester
 
 
-def test_gas_flag_in_tests(ape_pytester):
+def test_gas_flag_in_tests(ape_pytester, project):
     result = ape_pytester.runpytest("--gas")
     result.assert_outcomes(passed=NUM_TESTS), "\n".join(result.outlines)
